Renamed ast to cst.
I'm not happy with this verbiage of ast in it's current form, there
isn't enough semantic information to call this the actual ast of the
language in a form that expresses the programmer's intention.

The next stage will be to do some static analysis of this output cst and
product a (smaller) ast that expresses the true intentions of the
developer decorated with type information in it's almost final form
before moving onto the type engine to infer/deduce all final type
information and lifetimes of the program.