feat: Added a basic trie struct for keywords. Added span column support.
I'm using the trie to lookup symbols in the lexer at the moment but I'll
also use it for keyword lookup soon. It makes more sense to have a
library of tightly packed strings than huge, and difficult to read match
block.

This commit also adds an obvious thing I missed for diagnostic purposes,
column support in the spans which also respect Unicode widths.